# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface InstagramAnalysisInput {
  /**
     * Instagram username (with or without @) or a full profile URL
     * @minLength 1
     */
  input: string;
}

export interface InstagramProfile {
  username: string;
  fullName: string;
  profilePicUrl: string;
  biography: string;
  verified: boolean;
  followerCount: number;
  followingCount: number;
  postCount: number;
  externalUrl: string;
}

export interface InstagramPost {
  id: string;
  caption: string;
  /** Unix timestamp (seconds) the post was published */
  createTime: number;
  thumbnailUrl: string;
  postUrl: string;
  /** IMAGE, VIDEO, CAROUSEL_ALBUM, or REEL */
  mediaType: string;
  /** Number of likes */
  likeCount: number;
  commentCount: number;
  /** Number of views (for VIDEO/REEL only; 0 for IMAGE/CAROUSEL) */
  viewCount: number;
  /** Number of saves/bookmarks */
  saveCount: number;
  durationSeconds: number;
}

export interface InstagramAnalysisAggregates {
  totalPostsAnalyzed: number;
  totalLikeCount: number;
  totalCommentCount: number;
  /** Total views (VIDEO/REEL only) */
  totalViewCount: number;
  totalSaveCount: number;
  avgLikeCount: number;
  avgCommentCount: number;
  avgViewCount: number;
  avgSaveCount: number;
  /** Average (likes + comments + saves) / followers, as a percentage */
  engagementRate: number;
  postsPerWeek: number;
  /** @nullable */
  mostViewedPostId: string | null;
  /** @nullable */
  mostLikedPostId: string | null;
}

export interface PerformanceTierCounts {
  viral: number;
  tinggi: number;
  bagus: number;
  rataRata: number;
  rendah: number;
}

export interface HashtagStat {
  tag: string;
  count: number;
}

export interface MentionStat {
  mention: string;
  count: number;
}

export interface DayOfWeekStat {
  /** Indonesian day name (Senin..Minggu) */
  day: string;
  postCount: number;
  avgLikeCount: number;
  avgCommentCount: number;
  avgViewCount: number;
}

export interface MonthlyStat {
  /** Format YYYY-MM */
  month: string;
  postCount: number;
  totalLikeCount: number;
  avgLikeCount: number;
  avgEngagementRate: number;
}

export interface DurationBucketStat {
  bucket: string;
  postCount: number;
  avgViewCount: number;
  avgEngagementRate: number;
}

export interface YearlyStat {
  year: string;
  postCount: number;
  totalLikeCount: number;
  totalCommentCount: number;
  avgEngagementRate: number;
}

export interface MarketInsights {
  strengths: string[];
  weaknesses: string[];
  recommendations: string[];
}

export type IndustryBenchmarkEngagementRateComparison = typeof IndustryBenchmarkEngagementRateComparison[keyof typeof IndustryBenchmarkEngagementRateComparison];


export const IndustryBenchmarkEngagementRateComparison = {
  above: 'above',
  average: 'average',
  below: 'below',
} as const;

export type IndustryBenchmarkLikesPerFollowerComparison = typeof IndustryBenchmarkLikesPerFollowerComparison[keyof typeof IndustryBenchmarkLikesPerFollowerComparison];


export const IndustryBenchmarkLikesPerFollowerComparison = {
  above: 'above',
  average: 'average',
  below: 'below',
} as const;

export type IndustryBenchmarkPostingFrequencyComparison = typeof IndustryBenchmarkPostingFrequencyComparison[keyof typeof IndustryBenchmarkPostingFrequencyComparison];


export const IndustryBenchmarkPostingFrequencyComparison = {
  above: 'above',
  average: 'average',
  below: 'below',
} as const;

export interface IndustryBenchmark {
  engagementRateBenchmark: number;
  engagementRateComparison: IndustryBenchmarkEngagementRateComparison;
  likesPerFollowerBenchmark: number;
  accountLikesPerFollower: number;
  likesPerFollowerComparison: IndustryBenchmarkLikesPerFollowerComparison;
  postingFrequencyBenchmark: number;
  postingFrequencyComparison: IndustryBenchmarkPostingFrequencyComparison;
}

export type GrowthPotentialLabel = typeof GrowthPotentialLabel[keyof typeof GrowthPotentialLabel];


export const GrowthPotentialLabel = {
  rendah: 'rendah',
  sedang: 'sedang',
  tinggi: 'tinggi',
} as const;

export interface GrowthPotential {
  /**
     * @minimum 0
     * @maximum 100
     */
  score: number;
  label: GrowthPotentialLabel;
  reasoning: string;
}

export interface InstagramAnalysisInsights {
  topByViews: InstagramPost[];
  topByLikes: InstagramPost[];
  topByComments: InstagramPost[];
  performanceTiers: PerformanceTierCounts;
  topHashtags: HashtagStat[];
  topMentions: MentionStat[];
  performanceByDayOfWeek: DayOfWeekStat[];
  performanceByMonth: MonthlyStat[];
  durationAnalysis: DurationBucketStat[];
  yearlySummary: YearlyStat[];
  marketInsights: MarketInsights;
  industryBenchmark: IndustryBenchmark;
  growthPotential: GrowthPotential;
}

export interface AccountGrowthPoint {
  date: string;
  followerCount: number;
  postCount: number;
}

export interface InstagramAnalysis {
  id: number;
  input: string;
  profile: InstagramProfile;
  posts: InstagramPost[];
  aggregates: InstagramAnalysisAggregates;
  insights: InstagramAnalysisInsights;
  /** Follower count snapshots from every past analysis of this same account, oldest first */
  growthHistory: AccountGrowthPoint[];
  createdAt: string;
}

export interface InstagramAnalysisSummary {
  id: number;
  input: string;
  username: string;
  fullName: string;
  profilePicUrl: string;
  followerCount: number;
  postCount: number;
  createdAt: string;
}