# lib/api-zod/src/generated/types/dayOfWeekStat.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DayOfWeekStat {
  /** Indonesian day name (Senin..Minggu) */
  day: string;
  postCount: number;
  avgLikeCount: number;
  avgCommentCount: number;
  avgViewCount: number;
}