# lib/api-zod/src/generated/types/growthPotential.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GrowthPotentialLabel } from './growthPotentialLabel';

export interface GrowthPotential {
  /**
     * @minimum 0
     * @maximum 100
     */
  score: number;
  label: GrowthPotentialLabel;
  reasoning: string;
}