# lib/api-zod/src/generated/types/instagramAnalysis.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AccountGrowthPoint } from './accountGrowthPoint';
import type { InstagramAnalysisAggregates } from './instagramAnalysisAggregates';
import type { InstagramAnalysisInsights } from './instagramAnalysisInsights';
import type { InstagramPost } from './instagramPost';
import type { InstagramProfile } from './instagramProfile';

export interface InstagramAnalysis {
  id: number;
  input: string;
  profile: InstagramProfile;
  posts: InstagramPost[];
  aggregates: InstagramAnalysisAggregates;
  insights: InstagramAnalysisInsights;
  /** Follower count snapshots from every past analysis of this same account, oldest first */
  growthHistory: AccountGrowthPoint[];
  createdAt: Date;
}