# lib/api-zod/src/generated/types/instagramAnalysisAggregates.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InstagramAnalysisAggregates {
  totalPostsAnalyzed: number;
  totalLikeCount: number;
  totalCommentCount: number;
  /** Total views (VIDEO/REEL only) */
  totalViewCount: number;
  totalSaveCount: number;
  avgLikeCount: number;
  avgCommentCount: number;
  avgViewCount: number;
  avgSaveCount: number;
  /** Average (likes + comments + saves) / followers, as a percentage */
  engagementRate: number;
  postsPerWeek: number;
  /** @nullable */
  mostViewedPostId: string | null;
  /** @nullable */
  mostLikedPostId: string | null;
}