# lib/api-zod/src/generated/types/instagramAnalysisInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InstagramAnalysisInput {
  /**
     * Instagram username (with or without @) or a full profile URL
     * @minLength 1
     */
  input: string;
}