# lib/api-zod/src/generated/types/instagramAnalysisSummary.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InstagramAnalysisSummary {
  id: number;
  input: string;
  username: string;
  fullName: string;
  profilePicUrl: string;
  followerCount: number;
  postCount: number;
  createdAt: Date;
}