# lib/api-zod/src/generated/types/instagramPost.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InstagramPost {
  id: string;
  caption: string;
  /** Unix timestamp (seconds) the post was published */
  createTime: number;
  thumbnailUrl: string;
  postUrl: string;
  /** IMAGE, VIDEO, CAROUSEL_ALBUM, or REEL */
  mediaType: string;
  /** Number of likes */
  likeCount: number;
  commentCount: number;
  /** Number of views (for VIDEO/REEL only; 0 for IMAGE/CAROUSEL) */
  viewCount: number;
  /** Number of saves/bookmarks */
  saveCount: number;
  durationSeconds: number;
}