# lib/api-zod/src/generated/types/instagramProfile.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InstagramProfile {
  username: string;
  fullName: string;
  profilePicUrl: string;
  biography: string;
  verified: boolean;
  followerCount: number;
  followingCount: number;
  postCount: number;
  externalUrl: string;
}