# lib/api-zod/src/generated/types/marketInsights.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MarketInsights {
  strengths: string[];
  weaknesses: string[];
  recommendations: string[];
}