# lib/api-zod/src/generated/types/yearlyStat.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface YearlyStat {
  year: string;
  postCount: number;
  totalLikeCount: number;
  totalCommentCount: number;
  avgEngagementRate: number;
}